+++ app/src/main/java/dev/rikoapp/cleanphonelauncher/presentation/di/PresentationModule.kt
@@ -3,6 +3,7 @@ package dev.rikoapp.cleanphonelauncher.presentation.di
 import dev.rikoapp.cleanphonelauncher.presentation.applist.AppListViewModel
 import dev.rikoapp.cleanphonelauncher.presentation.home.HomeViewModel
 import dev.rikoapp.cleanphonelauncher.presentation.settings.SettingsViewModel
+import dev.rikoapp.cleanphonelauncher.presentation.widgets.WidgetsViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -10,4 +11,5 @@ val presentationModule = module {
     viewModelOf(::HomeViewModel)
     viewModelOf(::AppListViewModel)
     viewModelOf(::SettingsViewModel)
+    viewModelOf(::WidgetsViewModel)
 }